backend/requirements.txt
# Web Framework
Flask>=3.0.0
Flask-CORS>=4.0.0

# Deep Learning & Computer Vision
torch==2.9.1
torchvision==0.24.1
Pillow>=10.1.0

# NLP & RAG
langchain>=0.1.0
langchain-community>=0.0.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
numpy>=1.24.3
requests>=2.31.0
